% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canFuel_function.R
\name{canFuel}
\alias{canFuel}
\title{Canopy Fuel Stratum Characteristics Calculator}
\usage{
canFuel(ba, ht, tph, type)
}
\arguments{
\item{ba}{a vector or data frame with basal area [m2/ha]}

\item{ht}{a vector or data frame with average stand height [m]}

\item{tph}{a vector or data frame with stand density [trees/ha]}

\item{type}{a vector or data frame with forest cover type, either: Douglas-fir, Pseudotsuga menziesii,[DF]; ponderosa pine, Pinus ponderosa, [PP]; lodgepole pine, Pinus contorta, [LP]; or mixed conifer [MC]}
}
\value{
a data frame with: [1] canopy base height [m]; canopy fuel load [kg/m2]; and, canopy bulk density [kg/m3]
}
\description{
Estimation of canopy parameters using Cruz, Alexander & Wakimoto (2003).
}
\references{
Cruz M.G., Alexander M.E., Wakimoto R.H., 2003. Assessing canopy fuel stratum characteristics in crown fire prone fuel types of western North America. International Journal of Wildland Fire. 12(1):39-50.
}
\author{
Justin P Ziegler, \email{justin.ziegler@colostate.edu}
}
