% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfis_function.R
\name{cfis}
\alias{cfis}
\title{Canopy Fire Initiation & Spread model}
\usage{
cfis(fsg, u10, effm, sfc, cbd, id)
}
\arguments{
\item{fsg}{a vector or data frame of fuel stratum gap [m]}

\item{u10}{a vector or data frame with open (at a height 10 m above the canopy) wind speed [m/min]}

\item{effm}{a vector or data frame with stand density [trees/ha]}

\item{sfc}{a vector or data frame of surface fuel consumed [kg/m2]}

\item{cbd}{a vector or data frame of canopy bulk density [kg/m3]}

\item{id}{a vector or data frame of spot ignition delay, the time between which a firebrand is generated and the time it ignites on receptive fuels [min]}
}
\value{
a data frame with probability of crown fire occurences ["%"], type of fire, crown fire rate of spread [m/min], and critical spotting distance [m]
}
\description{
Estimation of crown fire initiation likelihood and crown fire rate of spread using Cruz, Alexander & Wakimoto (2004a,b) and .
}
\references{
Cruz M.G., Alexander M.E., Wakimoto R.H., 2004a. Modeling the likelihood of crown fire occurances in conifer forest stands. For. Sci. 50(4):640-658.
Development and testing of models for predicting crown fire rate of spread in conifer forest stands. Can. J. For. Res. 35:1626-1639.
Insert reference for spotting
}
\author{
Justin P Ziegler, \email{justin.ziegler@colostate.edu}
}
