% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coForest}
\alias{coForest}
\title{Colorado dry forest inventory summary.}
\format{A data frame with 14 observations of 10 variables:
\describe{
  \item{site}{name of forest location}
  \item{status}{either before (pre) or after (post) forest thinning}
  \item{trees_perha}{tree density (trees/ha)}
  \item{basalArea_m2ha}{basal area (m2/ha)}
  \item{qmd_cm}{quandratic mean diameter (cm)}
  \item{height_m}{mean tree height (m)}
  \item{sfl_kgm2}{surface fuel load (kg/m2)}
  \item{cbd_kgm3}{canopy bulk density (kg/m3)}
  \item{cbh_m}{canopy base height (m)}
  \item{cfl_kgm2}{canopy fuel load (kg/m2)}
}}
\source{
Ziegler, J.P., Hoffman, C., Battaglia, M., Mell, W., 2017. Spatially explicit measurements of forest structure
and fire behavior following restoration treatments in dry forests. \emph{Forest Ecology & Management}  \strong{386}, 1–12. doi:10.1016/j.foreco.2016.12.002
}
\usage{
data("coForest")
}
\description{
Fuels inventory summary of seven sampled forests in the southern Rocky Mountains and Colorado Plateau. Each forest was sampled before and after tree thinnings.
}
\keyword{datasets}
