% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fboTable}
\alias{fboTable}
\title{Fire Behavior Officer's table}
\format{A list with two data frames:
\describe{
  \item{1}{a list of reference fine fuel moistures (\%) by temperature (deg. C) and relative humidity (\%)}
  \item{2}{a list of correction fine fuel moistures (\%) by month (0-12), hourly time (0-23),
   shading, cardinal aspect, slope cateory, and elevation level category}
}}
\source{
Rothermel, R.C., Wilson, R.A., Morris, G.A., Sackett, S.S. 1986. Modeling moisture content of dead wildland fuels: input to the BEHAVE fire prediction system
 \emph{INT-RP-359}. US Department of Agriculture, Forest Service, Intermountain Forest and Range Experimental Station.
}
\usage{
data("fboTable")
}
\description{
Look up charts in tabular form to determine fine fuel moisture
}
\keyword{internal}
