% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fireChartData}
\alias{fireChartData}
\title{Template data for fire characteristics chart}
\format{An object of class \code{data.frame} with 9180 rows and 3 columns.}
\source{
Andrews, P.L., & Rothermel, R.C. 1982. Charts for interpreting wildland fire behavior characteristics.
 \emph{INT-GTR-131}. US Department of Agriculture, Forest Service, Intermountain Forest and Range Experimental Station.
}
\usage{
data("fireChartData")
}
\description{
Data of heat per unit area (kJ/m2), rate of spread (m/min), and flame length (m) for creating the template of the fire characteristics chart
}
\keyword{internal}
