% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{firefighter}
\alias{firefighter}
\title{Bitmap of firefighter
Bitmap image of bulldozer}
\format{An object of class \code{rastergrob} (inherits from \code{grob}, \code{gDesc}) of length 12.}
\usage{
firefighter
}
\description{
Bitmap of firefighter
Bitmap image of bulldozer
}
\seealso{
\code{\link{fireChart}}
}
\keyword{internal}
