% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fuelMoisture}
\alias{fuelMoisture}
\title{Modified Scott & Burgan (2005) moisture scenarios.}
\format{A data frame with 16 observations of 7 variables:
\describe{
  \item{fmLitter}{moisture of litter (\%)}
  \item{fm1hr}{moisture of 1-hr fuel (\%)}
  \item{fm10hr}{moisture of 10-hr fuel (\%)}
  \item{fm100hr}{moisture of 100-hr fuel (\%)}
  \item{fmLiveHerb}{moisture of herbaceous fuel (\%)}
  \item{fmLiveWoody}{moisture of woody fuel (\%)}
  \item{description}{scenario description}
}}
\source{
Scott, J., & Burgan, R. E. 2005. A new set of standard fire behavior fuel models for use with Rothermel’s surface fire spread model. \emph{RMRS-GTR-153}. Fort Collins, CO: US Department of Agriculture, Forest Service, Rocky Mountain Research Station.
}
\usage{
data(fuelMoisture)
}
\description{
Moisture scenarios are a set of fuel moistures of surface fuels, on a dry-weight basis, for each of the surface fuel classes. Originally developed by Scott & Burgan (2005), this dataset includes fuel moistures of litter.
}
\seealso{
\code{\link{nexus}}
}
\keyword{datasets}
