% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kbdiTable}
\alias{kbdiTable}
\title{KBDI Lookup table}
\format{An object of class \code{data.frame} with 1680 rows and 7 columns.}
\usage{
data("kbdiTable")
}
\description{
Data of drought factor by drought index, temperature and mean annual precipitation.
}
\keyword{internal}
