% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plume}
\alias{plume}
\title{Bitmap of fire plume
Bitmap image of fire plume}
\format{An object of class \code{rastergrob} (inherits from \code{grob}, \code{gDesc}) of length 12.}
\usage{
plume
}
\description{
Bitmap of fire plume
Bitmap image of fire plume
}
\seealso{
\code{\link{fireChart}}
}
\keyword{internal}
