% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rrRAWS}
\alias{rrRAWS}
\title{Rampart Range RAWS meteorological data}
\format{A data frame with 4392 observations of 4 variables:
\describe{
  \item{dateTime}{date and time of individual observation formatted as \code{"\%m/\%d/\%Y \%H:\%M"} }
  \item{temp_c}{air temperature (deg. C)}
  \item{rh}{relative humidity (\%)}
  \item{windSpeed_kmh}{wind speed (km/hr)}
  \item{precip_mm}{precipitation (mm)}
}}
\source{
RAWS USA Climate Archive \url{https://raws.dri.edu/}
}
\usage{
data("rrRAWS")
}
\description{
Hourly meteorological data from April 2017 through September 2017 from the Rampart Range Remote Automated Weather Station (RAWS; Station ID: RRAC2), maintained by the United States Forest Service.
}
\keyword{datasets}
