% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Bitmap of burning tree
Bitmap image of burning tree}
\format{An object of class \code{rastergrob} (inherits from \code{grob}, \code{gDesc}) of length 12.}
\usage{
tree
}
\description{
Bitmap of burning tree
Bitmap image of burning tree
}
\seealso{
\code{\link{fireChart}}
}
\keyword{internal}
