% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_function.R
\name{waf}
\alias{waf}
\title{Calculated wind adjustment factor}
\usage{
waf(fuelDepth, forestHt, cr, cc, sheltered = "n")
}
\arguments{
\item{fuelDepth}{a vector for depths of surface fuel bed (cm)}

\item{forestHt}{a vector of average stand tree heights (m)}

\item{cr}{a vector of crown ratios (\%)}

\item{cc}{a vector of canopy cover (\%)}

\item{sheltered}{a vector of either "y"es or "n"o as flags for using sheltered or unsheltered equations}
}
\value{
a vector of wind adjustment factors
}
\description{
Prediction of wind adjustment factor for sheltered and unsheltered fuels.
}
\details{
This calculates the wind adjustment factor (ratio of 10-m open wind speed to wind speed at the height of a surface fire).
One of two equations are used, depending on user input. By default, this function assumes the surface fuel bed is unsheltered.
\code{fuelDepth} must be a positive value if the unsheletered variant is invoked.
\code{forestHt} must be a positive value if the sheletered variant is invoked.
There are two conditions to enable calculation for a sheltered fuelbed. First, if  user could enter values for \code{cr} and
\code{cc} that lead to a crown fill portion above 5\%, or the user could enter \code{"y"} for \code{sheltered}. In the latter
case, this function assumes a crown fill portion of 10\%.
}
\examples{
library(firebehavioR)


}
\references{
Andrews, P.L. 2012. Modeling wind adjustment factor and midflame wind speed for Rothermel’s surface fire spread model. \emph{RMRS-GTR-266}. USDA Forest Service Rocky Mountain Research Station.
}
\author{
Justin P Ziegler, \email{justin.ziegler@colostate.edu}
}
